+++ pgark/__about__.py
@@ -1,8 +1,9 @@
-__title__ = 'pgark'
-__description__ = 'Python library and CLI for archiving URLs on popular services like Wayback Machine'
-__url__ = 'https://www.github.com/dannguyen/pgark'
-__version__ = '0.0.2'
-__author__ = 'Dan Nguyen'
-__author_email__ = 'dansonguyen@gmail.com'
-__license__ = 'MIT'
-
+__title__ = "pgark"
+__description__ = (
+    "Python library and CLI for archiving URLs on popular services like Wayback Machine"
+)
+__url__ = "https://www.github.com/dannguyen/pgark"
+__version__ = "0.0.2"
+__author__ = "Dan Nguyen"
+__author_email__ = "dansonguyen@gmail.com"
+__license__ = "MIT"