+++ pgark/__init__.py
@@ -2,5 +2,3 @@
 # -*- coding: utf-8 -*-
 
 from .__about__ import __version__
-
-