+++ pgark/cli.py
@@ -9,23 +9,37 @@ from pgark.mylog import mylogger
 from pgark.archivers import wayback
 
 OPTIONS_COMMON = [
-    click.option("-s", "--service", type=click.Choice(['wayback',]), default='wayback',
-        help="The service, e.g. wayback, permacc"),
-
+    click.option(
+        "-s",
+        "--service",
+        type=click.Choice(["wayback",]),
+        default="wayback",
+        help="The service, e.g. wayback, permacc",
+    ),
 ]
 
 OPTIONS_OUTPUT = [
-    click.option('-j', '--json', 'output_json', is_flag=True,
+    click.option(
+        "-j",
+        "--json",
+        "output_json",
+        is_flag=True,
         help="""By default, this subcommand returns a snapshot URL if successful, and nothing if not successful. Set this flag to return
-            the full JSON response"""),
-    click.option('-q', '--quiet', is_flag=True, help="Same as -v/--verbosity 0"),
-    click.option('-v', "--verbosity", type=click.IntRange(min=0, max=2), default=2,
+            the full JSON response""",
+    ),
+    click.option("-q", "--quiet", is_flag=True, help="Same as -v/--verbosity 0"),
+    click.option(
+        "-v",
+        "--verbosity",
+        type=click.IntRange(min=0, max=2),
+        default=2,
         help="""\b
                 Verbosity of log messages:
                   0: Silence (except errors)
                   1: Informational messages logged
                   2: Verbose debug log messages
-                  """),
+                  """,
+    ),
 ]
 
 
@@ -38,20 +52,22 @@ def _callback_print_version(ctx, param, value) -> NoReturn:
     cliprint(pgark.__version__)
     ctx.exit()
 
+
 def _set_verbosity(**kwargs):
     """TODO: should be decorator?"""
-    if kwargs.get('quiet') is True:
-        mylogger.setLevel('ERROR')
+    if kwargs.get("quiet") is True:
+        mylogger.setLevel("ERROR")
     else:
-        vb = kwargs.get('verbosity')
+        vb = kwargs.get("verbosity")
         if vb == 2:
-            mylogger.setLevel('DEBUG')
+            mylogger.setLevel("DEBUG")
         elif vb == 1:
-            mylogger.setLevel('INFO')
+            mylogger.setLevel("INFO")
         elif vb == 0:
-            mylogger.setLevel('ERROR')
+            mylogger.setLevel("ERROR")
 
-    mylogger.debug('Logger level: ', mylogger.get_level())
+    mylogger.debug("Logger level: ", mylogger.get_level())
+
 
 def add_options(*option_sets):
     def _decorate(func):
@@ -59,6 +75,7 @@ def add_options(*option_sets):
             for opt in reversed(options):
                 func = opt(func)
         return func
+
     return _decorate
 
 
@@ -68,14 +85,20 @@ def cliprint(obj) -> NoReturn:
     click.echo(obj)
 
 
-
 @click.group()
-@click.option("--version", callback=_callback_print_version, is_eager=True, is_flag=True, help='Print the version of pgark')
+@click.option(
+    "-v",
+    "--version",
+    callback=_callback_print_version,
+    is_eager=True,
+    is_flag=True,
+    help="Print the version of pgark",
+)
 def main(**kwargs):
     """
     Welcome to pgark
     """
-    if kwargs.get('version'):
+    if kwargs.get("version"):
         cliprint(pgark.__version__)
         return
 
@@ -89,16 +112,20 @@ def check(url, **kwargs):
     """
     _set_verbosity(**kwargs)
 
-
     answer, data = wayback.check_availability(url)
-    cliprint(data) if kwargs['output_json'] is True else cliprint(answer)
+    cliprint(data) if kwargs["output_json"] is True else cliprint(answer)
 
 
 @main.command()
 @add_options(OPTIONS_COMMON, OPTIONS_OUTPUT)
 @click.argument("url")
-@click.option("-u", "--user-agent", type=click.STRING, help="Specify a User-Agent header for the web request")
-def save(url, **kwargs):
+@click.option(
+    "-u",
+    "--user-agent",
+    type=click.STRING,
+    help="Specify a User-Agent header for the web request",
+)
+def save(url, user_agent, **kwargs):
     """
     Attempt to save a snapshot of [URL] using the [-s/--service]. Returns the snapshot URL if successful
     """
@@ -106,13 +133,18 @@ def save(url, **kwargs):
 
     snapshot_kwargs = {}
     if user_agent:
-        snapshot_kwargs['user_agent'] = user_agent
+        snapshot_kwargs["user_agent"] = user_agent
 
     answer, data = wayback.snapshot(url, **snapshot_kwargs)
-    if data['too_soon']:
-        mylogger.info(f"{data['too_soon_message']}", label='Wayback Machine response')
+    for i_key, i_value in data["issues"].items():
+        if i_value:
+            mylogger.info(f"{i_key}: {i_value}", label="Wayback Machine notice")
 
-    cliprint(data) if kwargs['output_json'] is True else cliprint(answer)
+    if kwargs["output_json"]:
+        cliprint(data)
+    else:
+        cliprint(answer)
+
 
 if __name__ == "__main__":
     main()