+++ pgark/exceptions.py
@@ -10,21 +10,27 @@ class CachedPage(Exception):
     This error is raised when archive.org declines to make a new capture
     and instead returns the cached version of most recent archive.
     """
+
     pass
 
 
-
-
-
-class SaveJobError(Exception):
+class SubmissionError(Exception):
     pass
 
+
+class AbnormalSubmitResponse(SubmissionError):
+    pass
+
+
+class SaveJobError(SubmissionError):
+    pass
 
 
 class WaybackRuntimeError(Exception):
     """
     An error returned by the Wayback Machine.
     """
+
     pass
 
 
@@ -32,7 +38,9 @@ class WaybackRobotsError(WaybackRuntimeError):
     """
     This error is raised when archive.org has been blocked by the site's robots.txt access control instructions.
     """
+
     pass
 
-class WaybackSubmitError(WaybackRuntimeError):
+
+class WaybackServerStatusError(WaybackRuntimeError):
     pass