+++ pgark/mylog.py
@@ -5,29 +5,23 @@ from rich.console import Console
 from rich.logging import RichHandler
 from rich.text import Text
 
-LOG_FORMAT = '%(message)s'
+LOG_FORMAT = "%(message)s"
 
 
 # uses the new markup API/param in Handler
 # https://github.com/willmcgugan/rich/issues/171
 HANDLERS = [
     RichHandler(
-        console=Console(file=stderr),
-        markup=True,
-        show_level=False,
-        show_path=False
+        console=Console(file=stderr), markup=True, show_level=False, show_path=False
     )
 ]
 logging.basicConfig(
-    datefmt="[%X]",
-    format=LOG_FORMAT,
-    handlers=HANDLERS,
-    level="DEBUG",
+    datefmt="[%X]", format=LOG_FORMAT, handlers=HANDLERS, level="DEBUG",
 )
 
 
 class MyLogger(object):
-    def __init__(self, loglib='rich', **kwargs):
+    def __init__(self, loglib="rich", **kwargs):
         self.logg = logging.getLogger(loglib)
 
     def get_level(self):
@@ -44,12 +38,13 @@ class MyLogger(object):
                 if len(args) == 1:
                     msg = args[0]
                 else:
-                    msg = ' '.join(str(a) for a in args)
+                    msg = " ".join(str(a) for a in args)
 
-                if kwargs.get('label'):
-                    label = '[purple on yellow]' +  kwargs.pop('label') + '[/]'
-                    msg = f'{label}: {msg}'
+                if kwargs.get("label"):
+                    label = "[purple on yellow]" + kwargs.pop("label") + "[/]"
+                    msg = f"{label}: {msg}"
                 decorated(cls, msg, **kwargs)
+
             return wrap
 
     @Decorators._fmt
@@ -71,7 +66,6 @@ class MyLogger(object):
     @Decorators._fmt
     def critical(self, *args, **kwargs):
         self.logg.critical(*args, **kwargs)
-
 
 
 mylogger = MyLogger()