+++ tests/test_wayback_subcommands.py
@@ -24,7 +24,7 @@ def test_check_success_and_available():
     resptext = EXAMPLES_DIR.joinpath("check/available-true.json").read_text()
 
     responses.add(
-        "GET", wb.availability_url(target_url), body=resptext,
+        "GET", wb.url_for_availability(target_url), body=resptext,
     )
 
     answer, data = wb.check_availability(target_url)
@@ -33,13 +33,15 @@ def test_check_success_and_available():
         answer
         == "http://web.archive.org/web/20200903230055/https://www.whitehouse.gov/issues/immigration/"
     )
+    assert data["target_url"] == target_url
     assert (
-        data["url"] == target_url
+        data["server_payload"]["url"] == target_url
     )  # TODO: not sure if this is guranteed, if target_url ends up being redirected??
-    df = data["archived_snapshots"]["closest"]
-    assert df["available"] is True
-    assert df["status"] == "200"
-    assert df["url"] == answer
+
+    ac = data["server_payload"]["archived_snapshots"]["closest"]
+    assert ac["available"] is True
+    assert ac["status"] == "200"
+    assert ac["url"] == answer
 
 
 @responses.activate
@@ -48,16 +50,18 @@ def test_check_success_but_not_available():
     resptext = EXAMPLES_DIR.joinpath("check/available-false.json").read_text()
 
     responses.add(
-        "GET", wb.availability_url(target_url), body=resptext,
+        "GET", wb.url_for_availability(target_url), body=resptext,
     )
 
     answer, data = wb.check_availability(target_url)
 
     assert answer is None
+    assert data["snapshot_url"] is None
+    assert data["target_url"] == target_url
     assert (
-        data["url"] == target_url
+        data["server_payload"]["url"] == target_url
     )  # TODO: not sure if this is guranteed, if target_url ends up being redirected??
-    assert data["archived_snapshots"] == {}
+    assert data["server_payload"]["archived_snapshots"] == {}
 
 
 @pytest.mark.skip(reason="TODO")
@@ -75,7 +79,7 @@ def test_check_success_not_ok():
 @responses.activate
 def test_save_job_polling():
     jid = "af709a09-c909-4883-b6b3-7350f9be8d7c"
-    url = wb.jobstatus_url(jid)
+    url = wb.url_for_jobstatus(jid)
     resptext = EXAMPLES_DIR.joinpath("job-status-success.json").read_text()
 
     responses.add("GET", url, body=resptext)
@@ -88,7 +92,7 @@ def test_save_job_polling():
 @responses.activate
 def test_snapshot_submit_request(session):
     target_url = "https://plainlanguage.gov/"
-    save_url = wb.savepage_url(target_url)
+    save_url = wb.url_for_savepage(target_url)
     resptext = EXAMPLES_DIR.joinpath(
         "job-save-success/submit-response.html"
     ).read_text()
@@ -114,7 +118,7 @@ def test_snapshot_submit_request(session):
 def test_snapshot_submit_request_not_ok(session):
     """not sure when this would happen, when server is down?"""
     target_url = "https://plainlanguage.gov/"
-    save_url = wb.savepage_url(target_url)
+    save_url = wb.url_for_savepage(target_url)
     resptext = EXAMPLES_DIR.joinpath(
         "job-save-success/submit-response.html"
     ).read_text()
@@ -130,7 +134,7 @@ def test_snapshot_submit_request_not_ok(session):
         ],
     )
 
-    with pytest.raises(WaybackSubmitError) as err:
+    with pytest.raises(WaybackServerStatusError) as err:
         resp = wb.submit_snapshot_request(session, target_url, headers={})
     assert (
         f"Server status was NOT OK; returned 503 for: {save_url}" in err.value.args[0]
@@ -145,11 +149,11 @@ def test_snapshot_successful():
     srcdir = EXAMPLES_DIR.joinpath("job-save-success")
 
     target_url = "https://plainlanguage.gov/"
-    save_url = wb.savepage_url(target_url)
+    save_url = wb.url_for_savepage(target_url)
 
     submit_resptext = srcdir.joinpath("submit-response.html").read_text()
     expected_job_id = wb.extract_job_id(submit_resptext)
-    expected_job_url = wb.jobstatus_url(expected_job_id)
+    expected_job_url = wb.url_for_jobstatus(expected_job_id)
 
     CURRENT_STATUS_ATTEMPT = 0
     status_paths = [
@@ -182,9 +186,7 @@ def test_snapshot_successful():
     )
 
     responses.add_callback(
-        "GET",
-        expected_job_url,
-        callback=_poll_callback,
+        "GET", expected_job_url, callback=_poll_callback,
     )
 
     answer, data = wb.snapshot(target_url, user_agent="guy incognito", poll_interval=0)
@@ -201,21 +203,24 @@ def test_snapshot_successful():
         answer
         == wb.BASE_DOMAIN
         + "/web/"
-        + data["last_job_status"]["timestamp"]
+        + data["server_payload"]["timestamp"]
         + "/"
         + target_url
     )
 
     # test data response
-    assert data["snapshot_status"] == "success"
+    assert data["was_new_snapshot_created"] is True
     assert data["snapshot_url"] == answer
     assert data["snapshot_request"]["user_agent"] == "guy incognito"
-    assert data["too_soon"] is False
-    assert data["too_soon_message"] == ""
+
+    issues = data["issues"]
+    assert issues["too_soon"] is False
+    assert issues["too_many_during_period"] is False
+
     assert data["job_id"] == expected_job_id
     assert data["job_url"] == expected_job_url
 
-    jd = data["last_job_status"]
+    jd = data["server_payload"]
     assert jd["status"] == "success"
     assert jd["timestamp"] in data["snapshot_url"]
 
@@ -226,13 +231,13 @@ def test_snapshot_successful():
 @responses.activate
 def test_snapshot_too_soon():
     srcdir = EXAMPLES_DIR.joinpath("job-save-too-soon")
-    target_url = 'https://plainlanguage.gov/'
+    target_url = "https://plainlanguage.gov/"
 
-    submit_resptext = srcdir.joinpath('submit-response.html').read_text()
+    submit_resptext = srcdir.joinpath("submit-response.html").read_text()
 
     responses.add(
         "POST",
-        wb.savepage_url(target_url),
+        wb.url_for_savepage(target_url),
         body=submit_resptext,
         status=200,
         match=[
@@ -244,19 +249,55 @@ def test_snapshot_too_soon():
 
     responses.add(
         "GET",
-        wb.jobstatus_url(wb.extract_job_id(submit_resptext)),
-        body=srcdir.joinpath('status-0.json').read_text(),
+        wb.url_for_jobstatus(wb.extract_job_id(submit_resptext)),
+        body=srcdir.joinpath("status-0.json").read_text(),
         status=200,
     )
 
     answer, data = wb.snapshot(target_url, poll_interval=0)
 
     assert answer == data["snapshot_url"]
-    assert data['snapshot_status'] == 'success'
-    assert data['too_soon'] is True
-    assert data['too_soon_message'] == 'The same snapshot had been made 4 minutes and 18 seconds ago. We only allow new captures of the same URL every 20 minutes.'
+    assert data["was_new_snapshot_created"] is False
 
-@pytest.mark.skip(reason="TODO")
+    assert (
+        data["issues"]["too_soon"]
+        == "The same snapshot had been made 4 minutes and 18 seconds ago. We only allow new captures of the same URL every 20 minutes."
+    )
+
+
+# @pytest.mark.skip(reason="Implement conditional branch after issue parsing")
 @responses.activate
-def test_snapshot_too_many():
-    pass
+def test_snapshot_too_many_for_period():
+    srcdir = EXAMPLES_DIR.joinpath("job-save-too-many-today")
+    submit_resptext = srcdir.joinpath("submit-response.html").read_text()
+    target_url = "https://nytimes.com/"
+
+    responses.add(
+        "POST",
+        wb.url_for_savepage(target_url),
+        body=submit_resptext,
+        status=200,
+        match=[
+            responses.urlencoded_params_matcher(
+                {"url": target_url, "capture_all": "on"}
+            )
+        ],
+    )
+    # mock request for availability URL
+    responses.add(
+        "GET",
+        wb.url_for_availability(target_url),
+        body=srcdir.joinpath("check-availability.json").read_text(),
+    )
+
+    answer, data = wb.snapshot(target_url, poll_interval=0)
+
+    assert answer == data["snapshot_url"]
+    assert data["was_new_snapshot_created"] == False
+    assert (
+        data["issues"]["too_many_during_period"]
+        == """This URL has been already captured 10 times today. Please email us at "info@archive.org" if you would like to discuss this more."""
+    )
+
+    # server payload ends up being availability API response
+    assert data["server_payload"]["archived_snapshots"]["closest"]["available"] is True